+++ app/main_dev/launch.js
@@ -194,8 +194,8 @@ export function closeDCRD() {
     dcrdRequest = null;
   } else if (isRunning(dcrdPID)) {
     try {
-      const win32ipc = require("win32ipc/build/Release/win32ipc.node");
-      win32ipc.closePipe(dcrdPipeRx);
+      const dcrwin32ipc = require("dcrwin32ipc/build/Release/dcrwin32ipc.node");
+      dcrwin32ipc.closePipe(dcrdPipeRx);
       dcrdPID = null;
       dcrdRequest = null;
     } catch (e) {
@@ -221,10 +221,10 @@ export const closeDCRW = () => {
       process.kill(dcrwPID, "SIGINT");
     } else if (isRunning(dcrwPID)) {
       try {
-        const win32ipc = require("win32ipc/build/Release/win32ipc.node");
+        const dcrwin32ipc = require("dcrwin32ipc/build/Release/dcrwin32ipc.node");
         dcrwTxStream.close();
-        win32ipc.closePipe(dcrwPipeTx);
-        win32ipc.closePipe(dcrwPipeRx);
+        dcrwin32ipc.closePipe(dcrwPipeTx);
+        dcrwin32ipc.closePipe(dcrwPipeRx);
       } catch (e) {
         logger.log("error", "Error closing dcrwallet piperx: " + e);
       }
@@ -249,8 +249,8 @@ export const closeDcrlnd = () => {
     dcrlndCreds = null;
   } else if (isRunning(dcrlndPID)) {
     try {
-      const win32ipc = require("win32ipc/build/Release/win32ipc.node");
-      win32ipc.closePipe(dcrlndPipeRx);
+      const dcrwin32ipc = require("dcrwin32ipc/build/Release/dcrwin32ipc.node");
+      dcrwin32ipc.closePipe(dcrlndPipeRx);
       dcrlndPID = null;
       dcrlndCreds = null;
     } catch (e) {
@@ -398,8 +398,8 @@ export const launchDCRD = (reactIPC, testnet, appdata) =>
 
     if (os.platform() == "win32") {
       try {
-        const win32ipc = require("win32ipc/build/Release/win32ipc.node");
-        dcrdPipeRx = win32ipc.createPipe("out");
+        const dcrwin32ipc = require("dcrwin32ipc/build/Release/dcrwin32ipc.node");
+        dcrdPipeRx = dcrwin32ipc.createPipe("out");
         args.push(format("--piperx=%d", dcrdPipeRx.readEnd));
       } catch (e) {
         logger.log("error", "can't find proper module to launch dcrd: " + e);
@@ -688,14 +688,14 @@ export const launchDCRWallet = async (
 
   if (os.platform() == "win32") {
     try {
-      const win32ipc = require("win32ipc/build/Release/win32ipc.node");
-      dcrwPipeRx = win32ipc.createPipe("out");
+      const dcrwin32ipc = require("dcrwin32ipc/build/Release/dcrwin32ipc.node");
+      dcrwPipeRx = dcrwin32ipc.createPipe("out");
       args.push(format("--piperx=%d", dcrwPipeRx.readEnd));
 
-      dcrwPipeTx = win32ipc.createPipe("in");
+      dcrwPipeTx = dcrwin32ipc.createPipe("in");
       args.push(format("--pipetx=%d", dcrwPipeTx.writeEnd));
       args.push("--rpclistenerevents");
-      const pipeTxReadFd = win32ipc.getPipeEndFd(dcrwPipeTx.readEnd);
+      const pipeTxReadFd = dcrwin32ipc.getPipeEndFd(dcrwPipeTx.readEnd);
       dcrwPipeTx.readEnd = -1; // -1 == INVALID_HANDLE_VALUE
 
       dcrwTxStream = fs.createReadStream("", { fd: pipeTxReadFd });
@@ -851,8 +851,8 @@ export const launchDCRLnd = (
 
     if (os.platform() == "win32") {
       try {
-        const win32ipc = require("win32ipc/build/Release/win32ipc.node");
-        dcrlndPipeRx = win32ipc.createPipe("out");
+        const dcrwin32ipc = require("dcrwin32ipc/build/Release/dcrwin32ipc.node");
+        dcrlndPipeRx = dcrwin32ipc.createPipe("out");
         args.push(format("--piperx=%d", dcrlndPipeRx.readEnd));
       } catch (e) {
         logger.log("error", "can't find proper module to launch dcrlnd: " + e);
