+++ modules/dcrwin32ipc/binding.gyp
@@ -1,7 +1,7 @@
 {
   "targets": [
     {
-      "target_name": "win32ipc",
+      "target_name": "dcrwin32ipc",
       "sources": [],
       "conditions": [
         ["OS=='win'", {"sources": ["module.cc", "pipe_wrapper.cc"]}]
