+++ modules/dcrwin32ipc/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "win32ipc",
+  "name": "dcrwin32ipc",
   "version": "0.1.0",
   "description": "Wraps anonymous pipe creation on Windows",
   "private": true,
