+++ modules/dcrwin32ipc/pipe_wrapper.h
